refactor: co-locate palette definitions, decouple theme.go from config
- Move DefaultPalette() from theme.go to themes_builtin.go alongside
  gruvboxDark(), so all palette definitions live in one file.

- Move ThemeNameFromConfig() from styles/theme.go to common/common.go,
  removing the config import from the styles package. The app package
  inlines the equivalent nil-guard since it cannot import common.

- Fix stray double-space in themes_builtin.go map literal.

theme.go is now a pure infrastructure file (types, validation, loading)
with no dependency on charmtone or config.
