+++ internal/ui/styles/themes_builtin.go
@@ -1,11 +1,62 @@
 package styles
 
+import "github.com/charmbracelet/x/exp/charmtone"
+
 // buildBuiltinThemes returns the complete map of built-in theme palettes.
 // Called once during package-level var initialization.
 func buildBuiltinThemes() map[string]ThemePalette {
 	return map[string]ThemePalette{
-		"charmtone":  DefaultPalette(),
+		"charmtone":    DefaultPalette(),
 		"gruvbox-dark": gruvboxDark(),
+	}
+}
+
+// DefaultPalette returns the default Crush theme palette using charmtone
+// colors from the upstream charmbracelet/x/exp/charmtone package.
+func DefaultPalette() ThemePalette {
+	return ThemePalette{
+		Name:   "Charmtone",
+		Author: "Charmbracelet",
+		Colors: ThemeColors{
+			Primary:   charmtone.Charple.Hex(),
+			Secondary: charmtone.Dolly.Hex(),
+			Tertiary:  charmtone.Bok.Hex(),
+
+			BgBase:        charmtone.Pepper.Hex(),
+			BgBaseLighter: charmtone.BBQ.Hex(),
+			BgSubtle:      charmtone.Charcoal.Hex(),
+			BgOverlay:     charmtone.Iron.Hex(),
+
+			FgBase:      charmtone.Ash.Hex(),
+			FgMuted:     charmtone.Squid.Hex(),
+			FgHalfMuted: charmtone.Smoke.Hex(),
+			FgSubtle:    charmtone.Oyster.Hex(),
+
+			Border:      charmtone.Charcoal.Hex(),
+			BorderFocus: charmtone.Charple.Hex(),
+
+			Error:   charmtone.Sriracha.Hex(),
+			Warning: charmtone.Zest.Hex(),
+			Info:    charmtone.Malibu.Hex(),
+
+			White:      charmtone.Butter.Hex(),
+			BlueLight:  charmtone.Sardine.Hex(),
+			Blue:       charmtone.Malibu.Hex(),
+			BlueDark:   charmtone.Damson.Hex(),
+			GreenLight: charmtone.Bok.Hex(),
+			Green:      charmtone.Julep.Hex(),
+			GreenDark:  charmtone.Guac.Hex(),
+			Red:        charmtone.Coral.Hex(),
+			RedDark:    charmtone.Sriracha.Hex(),
+			Yellow:     charmtone.Mustard.Hex(),
+
+			DiffInsertFg:      "#629657",
+			DiffInsertBg:      "#2b322a",
+			DiffInsertBgLight: "#323931",
+			DiffDeleteFg:      "#a45c59",
+			DiffDeleteBg:      "#312929",
+			DiffDeleteBgLight: "#383030",
+		},
 	}
 }
